requirements.txt
gradio~=3.38.0
loguru~=0.7.0
numpy~=1.25.1
pandas~=2.0.3
python-dotenv~=1.0.0
qdrant_client==1.3.1
safetensors~=0.3.1
sentence_transformers==2.2.2
SQLAlchemy~=2.0.19
tabulate==0.9.0
